Ollama_Local_MCP/requirements.txt
ollama
mcp
httpx
beautifulsoup4
lxml
ddgs
pymupdf
pillow
streamlit
python-dotenv
pytesseract
langchain
langchain-ollama
langchain-community
langchain-text-splitters
langgraph
langsmith
faiss-cpu
pytest
pytest-mock
respx
